% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/api.R
\name{hal_api}
\alias{hal_api}
\title{HAL API}
\usage{
hal_api()
}
\value{
An object of class \code{HALQuery}: a \code{\link{list}} containing at least the following
components:

\describe{
\item{\code{q}}{(default to "\verb{*:*}").}
\item{\code{fl}}{default to "\code{docid}" "\code{label_s}").}
\item{\code{fq}}{}
\item{\code{sort}}{}
\item{\code{rows}}{(default to \code{30}).}
\item{\code{start}}{(default to \code{0}).}
\item{\code{wt}}{(default to "\code{json}").}
}
}
\description{
Initialize a request to the HAL API.
}
\examples{
hal_api()
}
\concept{API}
