% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/facet.R
\docType{methods}
\name{hal_facet}
\alias{hal_facet}
\alias{hal_facet.HALQuery}
\title{Facet Search}
\usage{
hal_facet(x, ...)

\method{hal_facet}{HALQuery}(
  x,
  field,
  limit = 5,
  sort = c("index", "count"),
  prefix = NULL,
  contains = NULL,
  pivot = NULL,
  range = NULL,
  ignore_case = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{HALQuery} (typically returned by
\code{\link{hal_api}}).}

\item{...}{Currently not used.}

\item{field}{A \code{\link{character}} string specifying the field to group
by.}

\item{limit}{An \code{\link{integer}} giving the maximum number of results
per group.}

\item{sort}{A \code{\link{character}} string specifying the field to be
used to sort the results.}

\item{prefix}{A \code{\link{character}} string.}

\item{contains}{A \code{\link{character}} string.}

\item{pivot}{A \code{\link{character}} string.}

\item{range}{A \code{\link{list}} containing the following components:
"\code{range}", "\code{start}", "\code{end}", "\code{gap}".}

\item{ignore_case}{A \code{\link{logical}} scalar: should character case
be ignored?}
}
\value{
An object of class \code{\link[=hal_api]{HALQuery}}.
}
\description{
Facet Search
}
\examples{
\donttest{
library(magrittr)

## Most recent publication by journal
hal_api() \%>\%
  hal_query("archéologie") \%>\%
  hal_facet(
    field = "docType_s",
    limit = 100,
    sort = "count",
    range = list(
      range = "producedDateY_i",
      start = 2000,
      end = 2020,
      gap = 1
    )
  ) \%>\%
  hal_search(limit = 10)
}
}
\seealso{
Other query tools: 
\code{\link{hal_filter}()},
\code{\link{hal_group}()},
\code{\link{hal_query}()},
\code{\link{hal_select}()},
\code{\link{hal_sort}()}
}
\author{
N. Frerebeau
}
\concept{query tools}
