% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/query.R
\docType{methods}
\name{hal_filter}
\alias{hal_filter}
\alias{hal_filter.HALQuery}
\title{Filter Results}
\usage{
hal_filter(x, ...)

\method{hal_filter}{HALQuery}(x, value, field = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{HALQuery} (typically returned by \code{\link[=hal_api]{hal_api()}}).}

\item{...}{Currently not used.}

\item{value}{A \code{\link{character}} string specifying the value to be used to filter
the results.}

\item{field}{A \code{\link{character}} string specifying the field to filter along.}
}
\value{
An object of class \code{\link[=hal_api]{HALQuery}}.
}
\description{
Filter Results
}
\examples{
library(magrittr)

## Simple filer
hal1 <- hal_api() \%>\% hal_filter("file", "submitType_s")
hal1$fq

## Advanced filter
hal2 <- hal_api() \%>\% hal_filter(c("THESE", "HDR"), "docType_s")
hal2$fq

## Multiple filters
hal3 <- hal_api() \%>\%
  hal_filter("[NOW-1MONTHS/DAY TO NOW/HOUR]", "submittedDate_tdate") \%>\%
  hal_filter("-notice", "submitType_s")
hal3$fq

## Range filters
hal4 <- hal_api() \%>\% hal_filter("[2000 TO 2013]", "submittedDateY_i")
hal4$fq

hal5 <- hal_api() \%>\% hal_filter("[Aa TO Ab]", "city_s")
hal5$fq
}
\seealso{
Other query tools: 
\code{\link{hal_facet}()},
\code{\link{hal_group}()},
\code{\link{hal_query}()},
\code{\link{hal_select}()},
\code{\link{hal_sort}()}
}
\author{
N. Frerebeau
}
\concept{query tools}
