% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/group.R
\docType{methods}
\name{hal_group}
\alias{hal_group}
\alias{hal_group.HALQuery}
\title{Group Results}
\usage{
hal_group(x, ...)

\method{hal_group}{HALQuery}(x, field, limit = 1, sort = "score", decreasing = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{HALQuery} (typically returned by
\code{\link{hal_api}}).}

\item{...}{Currently not used.}

\item{field}{A \code{\link{character}} string specifying the field to group
by.}

\item{limit}{An \code{\link{integer}} giving the maximum number of results
per group.}

\item{sort}{A \code{\link{character}} string specifying the field to be
used to sort the results.}

\item{decreasing}{A \code{\link{logical}} scalar: should the sort be
increasing or decreasing?}
}
\value{
An object of class \code{\link[=hal_api]{HALQuery}}.
}
\description{
Group Results
}
\examples{
\donttest{
library(magrittr)

## Most recent publication by journal
hal_api() \%>\%
  hal_query("archéologie") \%>\%
  hal_select("producedDate_tdate") \%>\%
  hal_filter("ART" \%IN\% "docType_s") \%>\%
  hal_sort("producedDate_tdate", decreasing = TRUE) \%>\%
  hal_group(field = "journalTitle_s", limit = 1,
            sort = "producedDate_tdate", decreasing = TRUE) \%>\%
  hal_search(limit = 10)
}
}
\seealso{
Other query tools: 
\code{\link{hal_facet}()},
\code{\link{hal_filter}()},
\code{\link{hal_query}()},
\code{\link{hal_select}()},
\code{\link{hal_sort}()}
}
\author{
N. Frerebeau
}
\concept{query tools}
