% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/query.R
\docType{methods}
\name{hal_query}
\alias{hal_query}
\alias{hal_query.HALQuery}
\alias{\%OR\%}
\alias{\%AND\%}
\alias{\%NOT\%}
\alias{\%BY\%}
\alias{\%IN\%}
\alias{\%TO\%}
\title{Build Query}
\usage{
hal_query(x, ...)

\method{hal_query}{HALQuery}(x, value, field = NULL, ...)

x \%OR\% y

x \%AND\% y

x \%NOT\% y

x \%BY\% y

x \%IN\% y

x \%TO\% y
}
\arguments{
\item{x, y}{A \code{\link{character}} vector or an object of class
\code{HALQuery} (typically returned by \code{\link{hal_api}}).}

\item{...}{Currently not used.}

\item{value}{A \code{\link{vector}} giving the topics to be searched for
(see details).}

\item{field}{A \code{\link{character}} string specifying the field to search
within. If \code{NULL} (the default), search within "\code{text}".}
}
\value{
An object of class \code{\link[=hal_api]{HALQuery}}.
}
\description{
Build Query
}
\examples{
library(magrittr)

## Select fields
hal1 <- hal_api() \%>\% hal_query("asie")
hal1$q

term <- list("japon", "france")
hal2 <- hal_api() \%>\% hal_query(term, field = "title_t")
hal2$q

term <- list("Journal", c("Histoire", "History"))
hal3 <- hal_api() \%>\% hal_query(term, field = "title_t")
hal3$q

## Update query
hal4 <- hal_query(hal1, "agricol?")
hal4$q

## Operators
term <- list("Paris", "France", "history") \%NOT\% list("Texas", "history")
hal5 <- hal_api() \%>\% hal_query(term \%IN\% "text")
hal5$q

term <- c("aluminium", "fer") \%BY\% 3 \%IN\% "title_t"
hal6 <- hal_api() \%>\% hal_query(term)
hal6$q

term <- list("ecology" \%IN\% "title_t", "cell" \%IN\% "text")
hal7 <- hal_api() \%>\% hal_query(term)
hal7$q
}
\seealso{
Other query tools: 
\code{\link{hal_filter}()},
\code{\link{hal_group}()},
\code{\link{hal_select}()},
\code{\link{hal_sort}()}
}
\author{
N. Frerebeau
}
\concept{query tools}
