% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/query.R
\docType{methods}
\name{hal_select}
\alias{hal_select}
\alias{hal_select.HALQuery}
\title{Select Fields}
\usage{
hal_select(x, ...)

\method{hal_select}{HALQuery}(x, ...)
}
\arguments{
\item{x}{An object of class \code{HALQuery} (typically returned by
\code{\link{hal_api}}).}

\item{...}{One or more \code{\link{character}} string separated by commas
giving the fields to be returned.}
}
\value{
An object of class \code{\link[=hal_api]{HALQuery}}.
}
\description{
Select Fields
}
\examples{
library(magrittr)

## Select fields
hal1 <- hal_api() \%>\% hal_select("label_s")
hal1$fl

hal2 <- hal_api() \%>\% hal_select("halId_s", "uri_s")
hal2$fl

## Update query
hal3 <- hal_select(hal1, "docType_s")
hal3$fl
}
\seealso{
Other query tools: 
\code{\link{hal_facet}()},
\code{\link{hal_filter}()},
\code{\link{hal_group}()},
\code{\link{hal_query}()},
\code{\link{hal_sort}()}
}
\author{
N. Frerebeau
}
\concept{query tools}
