% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odyssey-package.R
\docType{package}
\name{odyssey-package}
\alias{odyssey-package}
\alias{_PACKAGE}
\alias{odyssey}
\title{odyssey: Access documents from the Open Archive HAL}
\description{
Access references and download documents from the Open Archive HAL 
    <https://hal.archives-ouvertes.fr/> using the search API.
}
\details{
\tabular{ll}{
\strong{Package:} \tab odyssey \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nfrerebeau/odyssey}
  \item \url{https://nfrerebeau.github.io/odyssey}
  \item Report bugs at \url{https://github.com/nfrerebeau/odyssey/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
