% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOdyssey.R
\name{runOdyssey}
\alias{runOdyssey}
\title{Run odyssey app function}
\usage{
runOdyssey()
}
\description{
Run odyssey app function
}
\examples{
This function runs the odyssey application.
odyssey()
}
\keyword{odyssey}
\keyword{shiny}

