% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/search.R
\name{search}
\alias{search}
\alias{hal_count}
\alias{hal_search}
\alias{hal_count.HALQuery}
\alias{hal_search.HALQuery}
\title{Search}
\usage{
hal_count(x, ...)

hal_search(x, ...)

\method{hal_count}{HALQuery}(x, path = "search", ...)

\method{hal_search}{HALQuery}(
  x,
  path = "search",
  limit = 30,
  start = 0,
  cursor = FALSE,
  type = "df",
  progress = getOption("odyssey.progress"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{HALQuery} (typically returned by
\code{\link{hal_api}}).}

\item{...}{Currently not used.}

\item{path}{A \code{\link{character}} string specifying the url path (see
\code{\link[solrium]{SolrClient}} and details).}

\item{limit}{An \code{\link{integer}} giving the maximum number of results.
According to HAL policy, it cannot exceed 10000.}

\item{start}{An \code{\link{integer}} specifying an absolute offset in the
complete sorted list of matches to be used as the beginning of the current
page. Only used if \code{cursor} is \code{FALSE}.}

\item{cursor}{A \code{\link{logical}} scalar: should a cursor be used for
the pagination of results? If \code{TRUE}, the \code{sort} parameter of the
query will set to "\code{docid asc}".}

\item{type}{A \code{\link{character}} string specifying the type of the
results. It must be one of "\code{df}" (default) or "\code{list}".}

\item{progress}{A \code{\link{logical}} scalar: should a progress bar for
for the request be printed? Only used if \R is being used interactively.}
}
\value{
\code{hal_count} returns the total number of results.

\code{hal_search} returns a \code{\link{data.frame}} or
\code{\link{list}} (according to \code{type}).
}
\description{
Search
}
\examples{
\dontrun{
library(magrittr)

## Simple search
topic <- list("archéologie", "Celtes", "France") # Combined with AND
hal_api() \%>\%
  hal_query(topic) \%>\%
  hal_search(limit = 10)

## Get a list of archaeological journals
topic <- c("archéologie", "archaeology", "archäologie") # Combined with OR
hal_api() \%>\%
  hal_query(topic) \%>\%
  hal_select("title_s", "issn_s") \%>\%
  hal_filter("" \%TO\% "*" \%IN\% "issn_s") \%>\%
  hal_sort("title_s") \%>\%
  hal_search(path = "ref/journal")

## Get the most recent archaeological publication (in French) by journal
hal_api() \%>\%
  hal_query("archéologie") \%>\%
  hal_select("producedDate_tdate") \%>\%
  hal_filter("ART" \%IN\% "docType_s") \%>\%
  hal_sort("producedDate_tdate", decreasing = TRUE) \%>\%
  hal_group(
    field = "journalTitle_s",
    sort = "producedDate_tdate",
    decreasing = TRUE
  ) \%>\%
  hal_search(limit = 10)

## Get a list of archaeological laboratories
## (only joint laboratories of the CNRS and a French university)
topic <- "archéologie" \%IN\% "text" \%AND\% "UMR" \%IN\% "code_t"
hal_api() \%>\%
  hal_query(topic) \%>\%
  hal_select("name_s", "acronym_s", "code_s") \%>\%
  hal_filter("VALID" \%IN\% "valid_s") \%>\%
  hal_sort("acronym_s", decreasing = TRUE) \%>\%
  hal_search(path = "ref/structure", limit = 15)
}
}
\references{
\href{https://api.archives-ouvertes.fr/docs/search}{HAL search documentation}.

\href{https://api.archives-ouvertes.fr/docs/ref}{HAL reference frame documentation}.
}
\seealso{
Other search tools: 
\code{\link{hal_download}()}
}
\author{
N. Frerebeau
}
\concept{search tools}
