% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/search.R
\name{search}
\alias{search}
\alias{hal_count}
\alias{hal_search}
\alias{hal_count.HALQuery}
\alias{hal_search.HALQuery}
\title{Search}
\usage{
hal_count(x, ...)

hal_search(x, ...)

\method{hal_count}{HALQuery}(x, path = "search", ...)

\method{hal_search}{HALQuery}(
  x,
  path = "search",
  limit = 30,
  start = 0,
  cursor = FALSE,
  parse = "df",
  concat = ",",
  progress = getOption("odyssey.progress"),
  verbose = getOption("odyssey.verbose"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{HALQuery} (typically returned by \code{\link[=hal_api]{hal_api()}}).}

\item{...}{Currently not used.}

\item{path}{A \code{\link{character}} string specifying the url path (see details).}

\item{limit}{An \code{\link{integer}} giving the maximum number of results.
According to HAL policy, it cannot exceed 10000.}

\item{start}{An \code{\link{integer}} specifying an absolute offset in the complete
sorted list of matches to be used as the beginning of the current page.
Only used if \code{cursor} is \code{FALSE}.}

\item{cursor}{A \code{\link{logical}} scalar: should a cursor be used for
the pagination of results? If \code{TRUE}, the \code{sort} parameter of the
query will set to "\code{docid asc}".}

\item{parse}{A \code{\link{character}} string specifying the type of the results.
It must be one of "\code{df}" (default) or "\code{list}" (see
\code{\link[solrium:SolrClient]{solrium::SolrClient()}}).}

\item{concat}{A \code{\link{character}} string specifying the character to
concatenate elements of longer than length \eqn{1}.}

\item{progress}{A \code{\link{logical}} scalar: should a progress bar for
for the request be printed? Only used if \R is being used interactively.}

\item{verbose}{A \code{\link{logical}} scalar: should extra information be reported?}
}
\value{
\itemize{
\item \code{hal_count()} returns the total number of results.
\item \code{hal_search()} returns a \code{\link{data.frame}} or \code{\link{list}} (according to \code{type}).
}
}
\description{
Search
}
\examples{
\donttest{
library(magrittr)

## Simple search
topic <- list("archéologie", "Celtes", "France") # Combined with AND
hal_api() \%>\%
  hal_query(topic) \%>\%
  hal_search(limit = 10)

## Get a list of archaeological journals
topic <- c("archéologie", "archaeology", "archäologie") # Combined with OR
hal_api() \%>\%
  hal_query(topic) \%>\%
  hal_select("title_s", "issn_s") \%>\%
  hal_filter("issn_s" \%IN\% "" \%TO\% "*") \%>\%
  hal_sort("title_s") \%>\%
  hal_search(path = "ref/journal")

## Get the most recent archaeological publication (in French) by journal
hal_api() \%>\%
  hal_query("archéologie") \%>\%
  hal_select("producedDate_tdate") \%>\%
  hal_filter("docType_s" \%IN\% "ART") \%>\%
  hal_sort("producedDate_tdate", decreasing = TRUE) \%>\%
  hal_group(
    field = "journalTitle_s",
    sort = "producedDate_tdate",
    decreasing = TRUE
  ) \%>\%
  hal_search(limit = 10)

## Get a list of archaeological laboratories
## (only joint laboratories of the CNRS and a French university)
topic <- "text" \%IN\% "archéologie" \%AND\% "code_t" \%IN\% "UMR"
hal_api() \%>\%
  hal_query(topic) \%>\%
  hal_select("name_s", "acronym_s", "code_s") \%>\%
  hal_filter("valid_s" \%IN\% "VALID") \%>\%
  hal_sort("acronym_s", decreasing = TRUE) \%>\%
  hal_search(path = "ref/structure", limit = 15)
}
}
\references{
\href{https://cwiki.apache.org/confluence/display/solr/Common+Query+Parameters}{Apache Solr documentation}.

\href{https://api.archives-ouvertes.fr/docs/search}{HAL search documentation}.

\href{https://api.archives-ouvertes.fr/docs/ref}{HAL reference frame documentation}.
}
\seealso{
Other search tools: 
\code{\link{hal_download}()}
}
\author{
N. Frerebeau
}
\concept{search tools}
